# Front End with code/Front End - React/ai-sdr-frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI SDR Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/prospect-discovery/accounts": {
      "post": {
        "tags": [
          "prospect-discovery"
        ],
        "summary": "Discover Accounts",
        "operationId": "discover_accounts_prospect_discovery_accounts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDiscoveryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prospect-discovery"
        ],
        "summary": "List Discovered Accounts",
        "operationId": "list_discovered_accounts_prospect_discovery_accounts_get",
        "parameters": [
          {
            "name": "icp_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icp Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Discovered Accounts Prospect Discovery Accounts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prospect-discovery/contacts": {
      "post": {
        "tags": [
          "prospect-discovery"
        ],
        "summary": "Discover Contacts",
        "operationId": "discover_contacts_prospect_discovery_contacts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactDiscoveryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDiscoveryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prospect-discovery"
        ],
        "summary": "List Discovered Contacts",
        "operationId": "list_discovered_contacts_prospect_discovery_contacts_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Discovered Contacts Prospect Discovery Contacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prospect-discovery/run": {
      "post": {
        "tags": [
          "prospect-discovery"
        ],
        "summary": "Run Discovery",
        "operationId": "run_discovery_prospect_discovery_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProspectDiscoveryRunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCriteria": {
        "properties": {
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries"
          },
          "geographies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Geographies"
          },
          "employee_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RangeInt"
              },
              {
                "type": "null"
              }
            ]
          },
          "revenue_range_million": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RangeFloat"
              },
              {
                "type": "null"
              }
            ]
          },
          "company_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Company Types"
          },
          "revenue_ranges": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Revenue Ranges"
          },
          "company_size_ranges": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Company Size Ranges"
          },
          "funding_stages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Funding Stages"
          },
          "tech_stack_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tech Stack Signals"
          }
        },
        "type": "object",
        "title": "AccountCriteria"
      },
      "AccountDiscoveryResult": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/DiscoveredAccount"
            },
            "type": "array",
            "title": "Accounts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "accounts",
          "total"
        ],
        "title": "AccountDiscoveryResult"
      },
...
      "AuditFields": {
        "properties": {
          "created_by": {
            "type": "string",
            "title": "Created By",
            "default": "system"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "change_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          }
        },
        "type": "object",
        "title": "AuditFields"
      },
...
      "ContactDiscoveryRequest": {
        "properties": {
          "account_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Account Ids"
          },
          "target_titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Titles"
          },
          "target_seniorities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Seniorities"
          }
        },
        "type": "object",
        "required": [
          "account_ids"
        ],
        "title": "ContactDiscoveryRequest"
      },
      "ContactDiscoveryResult": {
        "properties": {
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/DiscoveredContact"
            },
            "type": "array",
            "title": "Contacts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "contacts",
          "total"
        ],
        "title": "ContactDiscoveryResult"
      },
...
      "DiscoveredAccount": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "icp_id": {
            "type": "string",
            "title": "Icp Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url"
          },
          "industry": {
            "type": "string",
            "title": "Industry"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "employee_count": {
            "type": "integer",
            "title": "Employee Count"
          },
          "revenue_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Range"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          },
          "fit_score": {
            "type": "integer",
            "title": "Fit Score",
            "default": 0
          },
          "fit_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fit Reasons"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "new"
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Discovered At"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "icp_id",
          "company_name",
          "industry",
          "location",
          "employee_count"
        ],
        "title": "DiscoveredAccount"
      },
      "DiscoveredContact": {
        "properties": {
          "contact_id": {
            "type": "string",
            "title": "Contact Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "seniority": {
            "type": "string",
            "title": "Seniority"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email_verification_status": {
            "type": "string",
            "title": "Email Verification Status",
            "default": "unknown"
          },
          "phone_verification_status": {
            "type": "string",
            "title": "Phone Verification Status",
            "default": "unknown"
          },
          "contact_status": {
            "type": "string",
            "title": "Contact Status",
            "default": "active"
          },
          "is_former_employee": {
            "type": "boolean",
            "title": "Is Former Employee",
            "default": false
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence",
            "default": 0
          },
          "persona_match_score": {
            "type": "integer",
            "title": "Persona Match Score",
            "default": 0
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "new"
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Discovered At"
          }
        },
        "type": "object",
        "required": [
          "contact_id",
          "account_id",
          "full_name",
          "title",
          "department",
          "seniority"
        ],
        "title": "DiscoveredContact"
      },
      "DiscoveryRequest": {
        "properties": {
          "icp_definition": {
            "$ref": "#/components/schemas/ICPDefinition-Input"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "icp_definition"
        ],
        "title": "DiscoveryRequest"
      },
...
      "ExclusionCriteria": {
        "properties": {
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries"
          },
          "customer_statuses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Customer Statuses"
          },
          "company_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Company Names"
          },
          "employee_max_below": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Max Below"
          },
          "geographies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Geographies"
          }
        },
        "type": "object",
        "title": "ExclusionCriteria"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ICPDefinition-Input": {
        "properties": {
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "icp_id": {
            "type": "string",
            "title": "Icp Id"
          },
          "icp_name": {
            "type": "string",
            "title": "Icp Name"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "valid",
              "warning",
              "rejected"
            ],
            "title": "Status",
            "default": "valid"
          },
          "account_criteria": {
            "$ref": "#/components/schemas/AccountCriteria"
          },
          "persona_criteria": {
            "$ref": "#/components/schemas/PersonaCriteria"
          },
          "pain_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pain Points"
          },
          "offer_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Summary"
          },
          "exclusions": {
            "$ref": "#/components/schemas/ExclusionCriteria"
          },
          "scoring_weights": {
            "$ref": "#/components/schemas/ScoringWeights"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "audit": {
            "$ref": "#/components/schemas/AuditFields"
          },
          "audit_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Audit Metadata"
          }
        },
        "type": "object",
        "required": [
          "icp_name",
          "account_criteria",
          "persona_criteria",
          "scoring_weights"
        ],
        "title": "ICPDefinition"
      },
...
      "PersonaCriteria": {
        "properties": {
          "titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Titles"
          },
          "departments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Departments"
          },
          "seniorities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seniorities"
          },
          "seniority_levels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seniority Levels"
          },
          "buying_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Buying Signals"
          }
        },
        "type": "object",
        "title": "PersonaCriteria"
      },
...
      "ProspectDiscoveryRunResult": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/DiscoveredAccount"
            },
            "type": "array",
            "title": "Accounts"
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/DiscoveredContact"
            },
            "type": "array",
            "title": "Contacts"
          },
          "account_total": {
            "type": "integer",
            "title": "Account Total"
          },
          "contact_total": {
            "type": "integer",
            "title": "Contact Total"
          }
        },
        "type": "object",
        "required": [
          "accounts",
          "contacts",
          "account_total",
          "contact_total"
        ],
        "title": "ProspectDiscoveryRunResult"
      },
...
      "RangeFloat": {
        "properties": {
          "min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          }
        },
        "type": "object",
        "title": "RangeFloat"
      },
      "RangeInt": {
        "properties": {
          "min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          }
        },
        "type": "object",
        "title": "RangeInt"
      },
...
      "ScoringWeights": {
        "properties": {
          "industry_fit": {
            "type": "number",
            "title": "Industry Fit"
          },
          "company_size_fit": {
            "type": "number",
            "title": "Company Size Fit"
          },
          "persona_fit": {
            "type": "number",
            "title": "Persona Fit"
          },
          "geo_fit": {
            "type": "number",
            "title": "Geo Fit"
          },
          "pain_point_fit": {
            "type": "number",
            "title": "Pain Point Fit"
          }
        },
        "type": "object",
        "required": [
          "industry_fit",
          "company_size_fit",
          "persona_fit",
          "geo_fit",
          "pain_point_fit"
        ],
        "title": "ScoringWeights"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}